requirements_clean.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
SQLAlchemy==2.0.48
alembic==1.18.4
bcrypt==5.0.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.6
python-dotenv==1.0.1
qrcode==8.2
pillow==12.1.1
python-barcode==0.16.1
reportlab==4.2.5
weasyprint==61.2